Split out OCaml parser.
